feat: bf16オーディオ処理モジュールを追加しbf16対応を全体に拡充
- train/bf16_audio.py を新設（bf16でのDFT/mel等のオーディオ変換ユーティリティ）
- core（codec/model/rf/watermark 等）と inference/runtime を bf16 前提に更新
- DACVAEデコーダfinetuneスクリプトとコマンドを bf16 対応に整備
- train/optim.py に最適化まわりの拡張を追加
- bf16 推論・学習の検証テスト（tests/）を追加

+++ core/duration.py
@@ -164,7 +164,7 @@ def build_duration_features(
             ]
         )
 
-    return torch.tensor(rows, dtype=torch.float)
+    return torch.tensor(rows, dtype=torch.bfloat16)
 
 
 def set_duration_has_speaker_feature(
